requirements.txt
Flask==3.1.0
flask-restful==0.3.10
pymongo==4.6.1
python-dotenv==1.0.1